refactor!: classify TxSubmission errors as variant of CardanoNode error - remove lifecycle methods from CardanoNode interface, inherit them from RunnableModule - make TxSubmissionError a variant of CardanoNodeError - reuse Ogmios-related errors across CardanoNode and TxSubmitProvider - adjust src and tests with the new error namespace - generalize error names
