+++ packages/core/src/Cardano/util/txSubmissionErrors.ts
@@ -1,19 +1,19 @@
-import { TxSubmissionError, TxSubmissionErrors } from '../types/TxSubmissionErrors';
+import { CardanoNodeErrors } from '../../CardanoNode';
 
 /**
  * Tests the provided error for an instanceof match in the TxSubmissionErrors object
  *
- * @param {TxSubmissionError} error the error under test
+ * @param {CardanoNodeErrors.TxSubmissionError} error the error under test
  */
-export const isTxSubmissionError = (error: unknown): error is TxSubmissionError =>
-  Object.values(TxSubmissionErrors).some((TxSubmitError) => error instanceof TxSubmitError);
+export const isTxSubmissionError = (error: unknown): error is CardanoNodeErrors.TxSubmissionError =>
+  Object.values(CardanoNodeErrors.TxSubmissionErrors).some((TxSubmitError) => error instanceof TxSubmitError);
 
 /**
  * Attempts to convert the provided error or array of errors into TxSubmissionError object
  *
  * @param {any} error the error or array of errors under test
  */
-export const asTxSubmissionError = (error: unknown): TxSubmissionError | null => {
+export const asTxSubmissionError = (error: unknown): CardanoNodeErrors.TxSubmissionError | null => {
   if (Array.isArray(error)) {
     for (const err of error) {
       if (isTxSubmissionError(err)) {
