+++ packages/core/src/CardanoNode/types/CardanoNode.ts
@@ -1,6 +1,6 @@
 // eslint-disable-next-line @typescript-eslint/ban-ts-comment
 // @ts-ignore
-import { CardanoNodeError, CardanoNodeNotInitializedError } from './CardanoNodeErrors';
+import { CardanoNodeError } from './CardanoNodeErrors';
 import { HealthCheckResponse } from '../../Provider';
 import { Lovelace, PoolId, VrfVkHex } from '../../Cardano';
 
@@ -30,22 +30,6 @@ export type StakeDistribution = Map<
 >;
 
 export interface CardanoNode {
-  /**
-   * Initialize CardanoNode instance
-   */
-  initialize: () => Promise<void>;
-  /**
-   * Shut down CardanoNode instance
-   *
-   * @throws {CardanoNodeNotInitializedError}
-   */
-  shutdown: () => Promise<void>;
-  /**
-   * Get summaries of all Cardano eras
-   *
-   * @returns {EraSummary[]} Era summaries
-   * @throws {CardanoNodeError}
-   */
   eraSummaries: () => Promise<EraSummary[]>;
   /**
    * Get the start date of the network.
