+++ packages/core/src/Provider/TxSubmitProvider/types.ts
@@ -1,6 +1,6 @@
 // eslint-disable-next-line @typescript-eslint/ban-ts-comment
 // @ts-ignore
-import Cardano, { Provider } from '../..';
+import Cardano, { CardanoNodeErrors, Provider } from '../..';
 
 type SerializedTransaction = Cardano.util.HexBlob;
 
@@ -11,7 +11,7 @@ export interface SubmitTxArgs {
 export interface TxSubmitProvider extends Provider {
   /**
    * @param signedTransaction signed and serialized cbor
-   * @throws {Cardano.TxSubmissionError} (see Cardano.TxSubmissionErrors)
+   * @throws {CardanoNodeErrors.TxSubmissionError}
    */
   submitTx: (args: SubmitTxArgs) => Promise<void>;
 }
