+++ packages/core/test/Cardano/types/TxSubmissionErrors.test.ts
@@ -1,17 +1,16 @@
-import { Cardano } from '../../../src';
+import { CardanoNodeErrors } from '../../../src';
 import { TxSubmission } from '@cardano-ogmios/client';
-import { TxSubmissionErrors } from '../../../src/Cardano';
 
 describe('Cardano/types/TxSubmissionErrors', () => {
   test('TxSubmissionError can be narrowed down with "instanceof"', () => {
-    const error: Cardano.TxSubmissionError = new Cardano.TxSubmissionErrors.FeeTooSmallError({
+    const error: CardanoNodeErrors.TxSubmissionError = new CardanoNodeErrors.TxSubmissionErrors.FeeTooSmallError({
       feeTooSmall: { actualFee: 50n, requiredFee: 51n }
     });
-    expect(error).toBeInstanceOf(Cardano.TxSubmissionErrors.FeeTooSmallError);
+    expect(error).toBeInstanceOf(CardanoNodeErrors.TxSubmissionErrors.FeeTooSmallError);
   });
   test('maps all errors from Ogmios', () => {
     const ogmiosErrors = Object.keys(TxSubmission.submissionErrors.errors).sort();
-    const sdkErrors = Object.keys(TxSubmissionErrors)
+    const sdkErrors = Object.keys(CardanoNodeErrors.TxSubmissionErrors)
       .filter((e) => e !== 'UnknownTxSubmissionError')
       .map((e) => e.slice(0, Math.max(0, e.length - 5))) // drop 'Error' suffix
       .sort();
