+++ packages/ogmios/test/CardanoNode/OgmiosCardanoNode.test.ts
@@ -7,16 +7,15 @@ import { createMockOgmiosServer, listenPromise, serverClosePromise } from '../mo
 import { getRandomPort } from 'get-port-please';
 import { dummyLogger as logger } from 'ts-log';
 import http from 'http';
-
 describe('OgmiosCardanoNode', () => {
   let mockServer: http.Server;
   let connection: Connection;
   let node: OgmiosCardanoNode;
+  const moduleName = 'OgmiosCardanoNode';
 
   beforeAll(async () => {
     connection = createConnectionObject({ port: await getRandomPort() });
   });
-
   describe('not initialized and started', () => {
     beforeAll(async () => {
       mockServer = createMockOgmiosServer({
@@ -31,24 +30,23 @@ describe('OgmiosCardanoNode', () => {
 
     it('eraSummaries rejects with not initialized error', async () => {
       await expect(node.eraSummaries()).rejects.toThrowError(
-        new CardanoNodeErrors.CardanoNodeNotInitializedError('eraSummaries')
+        new CardanoNodeErrors.NotInitializedError('eraSummaries', moduleName)
       );
     });
     it('systemStart rejects with not initialized error', async () => {
       await expect(node.systemStart()).rejects.toThrowError(
-        new CardanoNodeErrors.CardanoNodeNotInitializedError('systemStart')
+        new CardanoNodeErrors.NotInitializedError('systemStart', moduleName)
       );
     });
     it('stakeDistribution rejects with not initialized error', async () => {
       await expect(node.stakeDistribution()).rejects.toThrowError(
-        new CardanoNodeErrors.CardanoNodeNotInitializedError('stakeDistribution')
+        new CardanoNodeErrors.NotInitializedError('stakeDistribution', moduleName)
       );
     });
     it('shutdown rejects with not initialized error', async () => {
       await expect(node.shutdown()).rejects.toThrowError(InvalidModuleState);
     });
   });
-
   describe('initialized and started', () => {
     describe('eraSummaries', () => {
       describe('success', () => {
@@ -65,13 +63,11 @@ describe('OgmiosCardanoNode', () => {
           await node.shutdown();
           await serverClosePromise(mockServer);
         });
-
         it('resolves if successful', async () => {
           const res = await node.eraSummaries();
           expect(res).toMatchSnapshot();
         });
       });
-
       describe('failure', () => {
         beforeAll(async () => {
           mockServer = createMockOgmiosServer({
@@ -97,7 +93,6 @@ describe('OgmiosCardanoNode', () => {
         });
       });
     });
-
     describe('systemStart', () => {
       describe('success', () => {
         beforeAll(async () => {
@@ -111,13 +106,11 @@ describe('OgmiosCardanoNode', () => {
           await node.shutdown();
           await serverClosePromise(mockServer);
         });
-
         it('resolves if successful', async () => {
           const res = await node.systemStart();
           expect(res).toMatchSnapshot();
         });
       });
-
       describe('failure', () => {
         beforeAll(async () => {
           mockServer = createMockOgmiosServer({
@@ -140,7 +133,6 @@ describe('OgmiosCardanoNode', () => {
         });
       });
     });
-
     describe('stakeDistribution', () => {
       describe('success', () => {
         beforeAll(async () => {
@@ -156,13 +148,11 @@ describe('OgmiosCardanoNode', () => {
           await node.shutdown();
           await serverClosePromise(mockServer);
         });
-
         it('resolves if successful', async () => {
           const res = await node.stakeDistribution();
           expect(res).toMatchSnapshot();
         });
       });
-
       describe('failure', () => {
         beforeAll(async () => {
           mockServer = createMockOgmiosServer({
@@ -188,30 +178,26 @@ describe('OgmiosCardanoNode', () => {
         });
       });
     });
-
     describe('shutdown', () => {
       beforeAll(async () => {
         mockServer = createMockOgmiosServer({ stateQuery: { systemStart: { response: { success: true } } } });
         await listenPromise(mockServer, connection.port);
       });
-
       afterAll(async () => {
         await serverClosePromise(mockServer);
       });
-
       beforeEach(async () => {
         node = new OgmiosCardanoNode(connection, logger);
         await node.initialize();
         await node.start();
       });
-
       it('shuts down successfully', async () => {
         await expect(node.shutdown()).resolves.not.toThrow();
       });
 
       it('throws when querying after shutting down', async () => {
         await node.shutdown();
-        await expect(node.systemStart()).rejects.toThrow(CardanoNodeErrors.CardanoNodeNotInitializedError);
+        await expect(node.systemStart()).rejects.toThrow(CardanoNodeErrors.NotInitializedError);
       });
     });
   });
