+++ packages/wallet/src/SingleAddressWallet/SingleAddressWallet.ts
@@ -10,6 +10,7 @@ import {
 import {
   AssetProvider,
   Cardano,
+  CardanoNodeErrors,
   ChainHistoryProvider,
   EpochInfo,
   EraSummary,
@@ -463,15 +464,15 @@ export class SingleAddressWallet implements ObservableWallet {
         // This could be improved by further parsing the error and:
         // - checking if ValueNotConservedError produced === 0 (all utxos invalid)
         // - check if UnknownOrIncompleteWithdrawalsError available withdrawal amount === wallet's reward acc balance
-        (error.innerError instanceof Cardano.TxSubmissionErrors.ValueNotConservedError ||
-          error.innerError instanceof Cardano.TxSubmissionErrors.UnknownOrIncompleteWithdrawalsError)
+        (error.innerError instanceof CardanoNodeErrors.TxSubmissionErrors.ValueNotConservedError ||
+          error.innerError instanceof CardanoNodeErrors.TxSubmissionErrors.UnknownOrIncompleteWithdrawalsError)
       ) {
         this.#logger.debug(`Transaction ${tx.id} appears to be already submitted...`);
         this.#newTransactions.pending$.next(tx);
         return;
       }
       this.#newTransactions.failedToSubmit$.next({
-        error: error as Cardano.TxSubmissionError,
+        error: error as CardanoNodeErrors.TxSubmissionError,
         reason: TransactionFailure.FailedToSubmit,
         tx
       });
